将鉴权移到分发服务器

+++ phantom-acceptor/src/main/java/com/phantom/acceptor/session/RedisSessionManager.java
@@ -1,8 +1,10 @@
 package com.phantom.acceptor.session;
 
+import com.alibaba.fastjson.JSONObject;
 import com.phantom.acceptor.config.AcceptorConfig;
 import com.phantom.common.Constants;
 import io.netty.buffer.Unpooled;
+import io.netty.channel.socket.SocketChannel;
 import lombok.extern.slf4j.Slf4j;
 import org.redisson.Redisson;
 import org.redisson.api.RBucket;
@@ -13,6 +15,11 @@ import org.redisson.client.codec.Codec;
 import org.redisson.client.protocol.Decoder;
 import org.redisson.client.protocol.Encoder;
 import org.redisson.config.Config;
+
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+import static com.phantom.common.Constants.SESSION_PREFIX;
 
 /**
  * 保存到redis的会话管理器
@@ -31,6 +38,16 @@ public class RedisSessionManager implements SessionManager {
      * 操作redis
      */
     private RedissonClient redissonClient;
+
+    /**
+     * 用户会话，表示接入系统内存中保存和用户的连接
+     */
+    private Map<String, SocketChannel> uid2Channel = new ConcurrentHashMap<>();
+
+    /**
+     * channel对应用户ID
+     */
+    private Map<SocketChannel, String> channel2Uid = new ConcurrentHashMap<>();
 
     private Decoder<Object> decoder = (buf, state) -> {
         if (buf == null) {
@@ -61,7 +78,7 @@ public class RedisSessionManager implements SessionManager {
     };
 
 
-    RedisSessionManager(AcceptorConfig acceptorConfig) {
+    public RedisSessionManager(AcceptorConfig acceptorConfig) {
         this.config = acceptorConfig;
         Config config = new Config();
         config.useClusterServers()
@@ -70,12 +87,39 @@ public class RedisSessionManager implements SessionManager {
         this.redissonClient = Redisson.create(config);
     }
 
+
     @Override
-    public void removeSession(String uid) {
-        String sessionKey = Constants.SESSION_PREFIX + uid;
-        RBucket<String> bucket = redissonClient.getBucket(sessionKey, codec);
-        log.info("从redis中删除session: {}", bucket.get());
-        bucket.delete();
+    public void addSession(String uid, Session session, SocketChannel channel) {
+        uid2Channel.put(uid, channel);
+        channel2Uid.put(channel, uid);
+        String key = SESSION_PREFIX + uid;
+        String value = JSONObject.toJSONString(session);
+        RBucket<Session> bucket = redissonClient.getBucket(key, codec);
+        bucket.set(session);
+        log.info("往redis中写入session ：{} -> {}", key, value);
+    }
+
+    @Override
+    public boolean removeSession(SocketChannel channel) {
+        String uid = channel2Uid.remove(channel);
+        if (uid != null) {
+            uid2Channel.remove(uid);
+            String sessionKey = Constants.SESSION_PREFIX + uid;
+            RBucket<String> bucket = redissonClient.getBucket(sessionKey, codec);
+            log.info("从redis中删除session: {}", bucket.get());
+            bucket.delete();
+        }
+        return uid != null;
+    }
+
+    @Override
+    public SocketChannel getChannel(String uid) {
+        return uid2Channel.get(uid);
+    }
+
+    @Override
+    public String getUid(SocketChannel channel) {
+        return channel2Uid.get(channel);
     }
 
 }
